Adding/removing methods of gen-tree op schedule.
Adding get_number_of_operators method, and removing
draw_operable_operator.

The latter messes up the sampling distribution. Seems obvious now, but
if you choose operators conditional on the state of the model (i.e., not
drawing operators that cannot operate on the current state), this
changes the sampling distribution (so it's not the posterior).