Updating gen tree operators.
Adding 'ignore_proposal_attempt_' boolean member that is set to true
when the proposal returns -inf because it cannot operate on the current
state of the model (and thus never attempted to make a move).  When this
is true, the move is not counted, and no autotuning is performed.

Now that I know that I cannot only draw operators that can operate on
the current state (it's obvious now that this can bias the sampling
distribution), this is an alternative strategy. We don't want the
operator not being able to even try a move to count as a rejection and
contributed to autotuning.

Also, fixing bug (I think) in NeighborHeightNodePermute and
NeighborHeightNodeSwap. These were trying swap/permute moves on height
index i and i + 1. This is wrong, because i + 1 may very well not have
any parent nodes of i. Fixing this to try move on i and 'youngest index
of height with parent node of i'. These fixes still need to be tested.