Working on tests of gen tree rev jump move.
Working on "full model" tests of gen tree rev jump move.

The test that constrains pop sizes to be equal appears to be working.

The test that allows each branch to have a free pop size parameter is
failing, because currently, the Hasting's ratio of the rev jump move
does not handle the probas associated with adding/removing pop size
parameters (only node heights). At least, I assume that is why it is
failing.