Working on RJ gen tree move with free pop sizes.
Trying to get the SplitLumpNodesRevJumpSampler operator working when the
population sizes are constrained to be equal across the whole tree
(i.e., each branch gets its own free pop size parameter).

I think this was failing, because the RJ move did not consider the
probability of adding/removing new pop size parameters to the model.
Trying to remedy this by randomly drawing pop sizes, and incorporating
the probability of these draws into the Hasting's ratio of of the RJ
move.

So far, this is NOT working. The test with free pop size parmeters is
still failing badly.