Working on PositiveRealParameterSettings class.
Adding boolean option to init_from_yaml_node so that we can create
instances that are fixed without a value. This will be needed for the
root node height when the user provides a tree (the user should be able
to fix the root node to the value in their tree, but we don't know that
value until after the settings class instance is created).

Also, adding set_value function. Do to above, we need to be able to set
the value later.