Reworking TreePriorSettings class and descendants.
Because the main settings instance will only have a pointer for the base
of the TreePriorSettings class hierarchy, we need a universal way to get
our hands on the parameter settings.
Doing this by adding a get_parameter_settings function that returns
a map to copies of the parameter settings.

Also, adding more sanity checks to TreeModelSettings to make sure the
various tree model settings are compatible.