Updating gen tree operatiors.
Adding set_default_weight to BaseGeneralTreeOperatorTemplate to allow
operator classes to set sane default weights as a function of the number
of leaves in the tree.

Also, adding helper_ops member and operate_with_helpers function to
GeneralTreeOperatorTemplate. These allow a vector of helper operators to
be associated with a op instance; these get called after the op performs
a move (when operate_with_helpers is called). This offers an alternative
to having to supply helper ops "on the fly" with the operate_plus
function.

Tests in test_general_tree_operator.cpp are running and none have failed
so far, so I don't think I broke anything.