Working on phycoevolity interface and mcmc.
Adding variable for the number of moves to perform each MCMC generation,
and making this an argument for the mcmc function (which still needs to
be written). Very much a work-in-progress.

In order for MCMC generation settings to scale as the size of the
problem (tree) grows, I was going to have different types of operators
perform more or less moves (the number could be set as a function of the
number of leaves, for example) each time they were selected during MCMC.
While flexible, this was messy. I'm changing this strategy so that the
new variable for the number of moves to perform each MCMC generation
will be solely responsible for allowing MCMC settings to scale with tree
size. This allows the operator weights to be specifically for
controlling how often they should be called relative to other operators.