Fixinb bug and adding constructor to op classes.
Adding constuctor to GeneralTreeOperatorInterface that takes
both the operator's weight and tuning parameter.

Adding this constructor to all descendants of
GeneralTreeOperatorInterface that have a tunable op member.

Also, fixing bug in GlobalNodeHeightDirichletOperator. It was not making
the tree 'dirty' and so would sample from the prior rather than the
posterior when data were not ignored (because the likelihood would not
be recalculated and thus would be the same).