Updating BaseTree::collision_node_permute.
Adding rejection sampling to make sure a change in topology happens.

Updating associated tests in test_tree.cpp. All of these updated test
are passing, along with the tests of the general tree operators that use
this function.

This will prevent the NeighborHeightNodePermute operator from wasting
computation making proposals that don't change the state of the model.