Updating test in test_simphycoeval_cli.cpp.
The test suite has been running for a long time, and so far, only one
test in test_simphycoeval_cli.cpp has failed due to Monte Carlo error.
This commit increases the number of samples collected for that test, and
it is now passing.

I can't say that ALL tests are passing, given the test suite is still
running, but I can say that at least the vast majority of tests are
passing.