Updating node classes.
Adding 'get_parameter_map' function to BaseNode, and overriding it in
PopulationNode. For BaseNode, this function receives a map, but does
nothing.  PopulationNode overrides this to put the population size into
the map.

Also, updating 'extract_data_from_node_comments' function for BaseNode
and Node so that the comment map is provided as a const reference
(rather than copied). All tests in 'test_tree_io.cpp' are passing so
this change didn't seem to break anything.