Updating NodeHeightSlideBump*Mover tests.
One test of NodeHeightSlideBumpSwapMover was failing on the
ASC cluster. Adding additional with NodeHeightSlideBumpSwapMover
operators for this test with varying window widths.

For other tests, making sure the root height starting value matches
the mean of the prior on the root height.

All of these tests are passing. Given this was the only failure in the
entire test suite, all tests in the suite are now passing!