Working on treesum tests.
Adding test of sample of heights. Tests now confirm that the exptected
behavior is working in terms of the counts of splits, heights, and
topologies.

More tests needed to confirm the parameter values are being stored
properly. More functionality of treesum classes needed to actually
summarize the samples of splits, heights, trees, and parameter values.