Working on treesum.
Adding class for summarizing the number of heights across sampled trees.
Making this new class a member of TreeSample.  Adding functions to
TreeSample to get the count/freq for a number of heights.

Also, adding funcitons to TreeSample to get the credibility level
of a topology and number of heights.

Adding functions to TreeSample to write MAP and target trees
to nexus format.

Adding tests of new stuff to test_treesum.cpp. All tests
in test_treesum.cpp are passing.