Updating TreeSample summary output
Updating the summary output of TreeSample to handle situations when the
target tree has heights and/or splits that were not sampled.

This was causing a segfault, because we were trying to look up the
height or split in a map, but it wasn't there.

When writing the summary of the target tree heights and nontrivial
splits, there are now checks to see if the height/split was sampled, and
if not, the output summary stats are "nan".

All tests in test_treesum.cpp and test_sumphycoeval_cli.cpp are passing.
No other tests are affected by the changes.