Adding root relationship info to treesum.
Updating treesum::TreeSample to keep track of and summarize the rooting
relationships of sampled trees.

Updating BaseTree::store_splits_heights_parameters function to store the
set of splits that descend from the root. This is needed for the
treesum::TreeSample to record root relationships.

Adding tests to test_treesum.cpp of the new behavior. All tests in
test_treesum.cpp and test_sumphycoeval_cli.cpp are passing, and these
should be the only tests affected by the new changes.