Reworking treesum::TreeSample to store node info.
Previously, we were storing splits and heights, but with polytomies
allowed, splits are not as useful. We often want to know the
configuration of the splits that descend from a node, so we can report
things like the post prob of a polytomy.

Working on summarizing this info from a sample of trees. Note, this
makes the "root descendants" info we were tracking unnecessary, because
we will now be summarizing this info for all nodes (not just the root).

This is a work-in-progress and things are likely broken at the moment.
In fact, I know they are, because I am changing the "taxonomy" of the
summary output to help make the "split"/"node" distinction more clear,
and I haven't updated the tests to match this new language.