Updating TopologySamples class to store node info.
Storing a split-to-node map and node-to-split map for each instance of
TopologySamples. This allows us to report node information when
summarizing a topology.

Updating summary output functions of TreeSample to output node info
for all topologies.

This required updating the BaseTree::store_splits_heights_parameters
function to populate a reference to a split-to-node map.

Updating tests for the new behavior.

All tests in test_split.cpp, test_treesum.cpp, test_treecomp.cpp, and
test_sumphycoeval_cli.cpp are passing.