Fixing bug affecting simphycoeval.
Simphycoeval was crashing when trying to simulate data under an
(unfixed) generalized model. The reason for this is that the internal
node indices were not unique. The split RJ MCMC move adds new nodes to
the tree, and was giving them an arbitrary index of -1. This was fine
for inference (i.e., sampling from the posterior), because only the
indices of the leaf nodes matter. However, the 'simulate_gene_tree'
member function of BasePopulationTree assumes all internal nodes have
unique indices, so simulating a character on a tree that had split RJ
MCMC moves performed on it would crash.

In this commit, I'm adding 'BaseTree.update_internal_node_indices' to
keep internal node indices positive and unique. This is called whenever
the tree is initiated (with 'set_root'), whenever a node is added or
removed from the tree, and whenever 'restore_topology' is called.

This fixes the simphycoeval crash. All tests in:

    test_phycoeval_cli.cpp
    test_simphycoeval_cli.cpp
    test_tree.cpp
    test_tree_io.cpp
    test_stats_util.cpp
    test_treecomp.cpp
    test_treesum.cpp
    test_sumphycoeval_cli.cpp
    test_split.cpp

are passing. I didn't run the rest of the test suite.