Changing proposal dist of gen tree RJ MCMC op.
Changing the proposal distribution of the SplitLumpNodesRevJumpSampler
operator from uniform to beta. When splitting a height down (tipward),
this is the distribution the new height is drawn from.

A beta will provide more flexibility in how different the new proposed
height is from form the original height. Currently, this is hard-coded
to a beta(1,1) and so it doesn't take advantage of this flexibility.

I've only done minimal testing, but the switch to a beta seems to be
working.