Fixing bug in simphycoeval.
When writing configs for analyzing simulated data, simphycoeval was
passing the operator schedule of the simulation settings to
'write_settings'. This replaced the intended MCMC analysis settings in
the prior configs.

Fixing this by creating operator schedules for each prior
config/settings and using them when calling write_settings.

All tests in:

    test_general_tree_settings.cpp
    test_general_tree_operator_schedule.cpp
    test_simphycoeval_cli.cpp
    test_tree_io.cpp
    test_treecomp.cpp
    test_treesum.cpp
    test_sumphycoeval_cli.cpp
    test_split.cpp

are passing.