Adding/updating gen tree RJ op tests.
After changing the SplitLumpNodesRevJumpSampler to use a beta
proposal distribution, one of the 3-leaf tests in
test_general_tree_rj_operator.cpp was failing, apparently due to Monte
Carlo error (draws from beta(1,1) will not produce identical
results as the old uniform proposal distribution, even with the same
seed). Increasing the number of samples for this test, and it is now
passing.

Adding a 5-leaf test of the new auto optimizing behavior of
SplitLumpNodesRevJumpSampler. This is passing.

I have not run all the tests in test_general_tree_rj_operator.cpp, but
these two and some others I've run are passing.