Reworking handling of operator settings.
Overhauling how 'general_tree_settings.hpp' handles operators to allow
multiple operators of the same type to be specified.
Added some convenience functions to the operators themselves in
'general_tree_operator.hpp' to make this easier.
This also required updates to 'general_tree_operator_schedule.hpp'
and 'test_general_tree_operator_schedule.cpp'.

The purpose of all this is to use multiple SplitLumpNodesRevJumpSampler
operators with different alphas for their beta proposal distribution
(and allow these to be optimized).

Adding test to 'test_general_tree_rj_operator.cpp' to check that
multiple SplitLumpNodesRevJumpSampler operators with different alphas,
and with one optimizing, samples from the expected prior distribution.
This test is passing.

Also, all tests in:

    test_general_tree_settings.cpp
    test_general_tree_operator_schedule.cpp
    test_phycoeval_cli.cpp
    test_simphycoeval_cli.cpp
    test_tree_io.cpp
    test_stats_util.cpp
    test_treecomp.cpp
    test_treesum.cpp
    test_sumphycoeval_cli.cpp
    test_split.cpp

are passing. And the cmake build is working.