Fixing newick bug in treesum.
Fixing bug in the output of newick trees (from splits) in treesum.hpp,
which resulted in a crash and error message from
HeightSamples::reverse_sort_by_nesting.
Basically the logic in this sorting function was causing children to
sometimes come before their parents in some cases of shared node
heights.
Updating HeightSamples::reverse_sort_by_nesting to fix this.
This bug only affected the writing of newick trees by the sumphycoeval
tool, and resulted in an error (not a bogus tree).

All tests in

-   test_treesum.cpp
-   test_sumphycoeval_cli.cpp

are passing.