Overhauling how TreeSample outputs newick trees.
TreeSample::to_parentheses was trying to use sets of splits to build a
tree and write it in newick format. This was fragile, and shared heights
among nodes could break it.

Overhauling how this is done. We now store descendant splits of each
node in TopologySamples, and use this information to robustly build
trees, annotate nodes, and write to newick/nexus format.

Also, when writing the target tree, we are now using the target's node
heights, rather than sample mean/median. This avoids the problem of
having to deal with unsampled nodes/heights.

All tests in:
-   test_treesum.cpp
-   test_sumphycoeval_cli.cpp
are passing.