Adding test_phylonet.cpp.
Most of these tests are copied and updated from test_tree.cpp.

One test is new (not from test_tree.cpp), and compares the frequency of
simulated patterns to likelihood values. The scenario is simple; one
population with 2 alleles, so there are three possible site patterns:
2,0; 2,1; 2,2.
This test is showing odd results. The sims and likelihoods match for 2,0
and 2,2, but the likelihood of 2,1 is half the frequency of that pattern
in the sims.
Cranking up the pop size (coalescent time) confirms the sims are
correct. We see:

        Sim_freq    Likelihood
2,0     ~ 0.25      ~ 0.25
2,1     ~ 0.5       ~ 0.25
2,2     ~ 0.25      ~ 0.25

All tests in test_phylonet.cpp are passing, but for the test discussed
above, it checks that the likelihood of 2,1 is half of the sim
frequency, which is unexpected.