Working on tree and phylonet tests.
Working on and adding tests that compare likelihoods of patterns to
their frequencies from simulations. I figured out why they are
different. The likelihood is corrected for all the ways you can get the
same allele count pattern. When you "uncorrect" it, it matches the
simulations, which do not distinguish among the possibilites when
approximating the pattern frequency.