More cleaning up after net node overhaul.
Adding functions to net node classes (and phylonet class) to specify the
branch of a node with multiple parents by providing the parent the
branch leads to (rather than the index of the parent).  This simplified
their use in the netlikelihood functions (which needed a lot of cleaning
up to match the new attributes/behavior of net node classes.