Adding test to test_netlikelihood.cpp.
This test confirms that the split_top_of_branch_partials function works
the way it should when the probability of the alleles going to two
parents are 1/0.
This suggests that my math is wrong, rather than the function having
bugs.

Currently, when the probability of alleles going to a parent is zero,
I expect the conditional probability of the 0,0 pattern at the bottom
of that parent should be 1.0. However, the way I currently calculate
this assumes the conditional probs at the top of the daughter
sum to 1.0, but they don't have to, of course.