Adding node heights to treesum.
Shared (and singleton) heights were being summarized using sets of
splits for the shared divergences. Here, I'm adding the summarizing of
shared (and singleton) heights using sets of nodes, and each node is a
set of splits that descend from it.

Adding NodeHeightSamples class to store this new info. Adding a vector
and map of these NodeHeightSamples as members to TreeSample class.

Adding write_summary_of_merged_target_heights method to TreeSample.
This function reports the post probs of shared node heights created by
merging neighboring heights in the target tree. Storing heights as sets
of nodes (as opposed to splits) was necessary for this to work sanely.
For example, otherwise merging the penultimate height with the root
resulted in a post prob of 1, because all trees have the root "split".

Adding tests of this new functionality to test_treesum.cpp.  All tests
in test_treesum.cpp are passing.