Adding option to sumphycoeval.
Adding '--include-merged-target-heights' option to sumphycoeval.  This
takes advantage of the new functionality in treesum to report the post
prob of shared/multifurcating nodes created by merging neighboring
heights in the target tree.

The primary purpose for this option will be for simulations where the
target tree is the true tree.