Fixing bug in write_summary_of_merged_target_heights
Fixing bug in new function of TreeSample called
write_summary_of_merged_target_heights.
I thought the `number_of_resulting_merged_nodes` recorded during the
BaseTree::merge_node_height_up function was the total number of nodes
mapped to the height after the merge, but it is only the number of nodes
after the merge that involve nodes mapped to the younger height.
This was causing an assertion (to make sure the number of merged nodes
was correct) to fail.
Fixing this by explicitly getting the number of merged nodes after the
merge.

Adding new test to test_treesum.cpp to check for this problem.

All tests in test_treesum.cpp and test_sumphycoeval_cli.cpp
are passing.