Adding test confirming disjoint site labels.
Adding a test that demonstrates that the sequence labels can be
different at every site in a nexus alignment, and the resulting
BiallelicData will be identical.

This is relevant when concatenating the alignments of  multiple loci
together and the sequence labels do not match among the alignments.
This will result in the loci being "disjoint" in the resulting
concatenated alignment, but this will not change how BiallelicData will
see the data.

For example, after concatenating two alignments together, you might end
up with a character matrix that looks like this if the sequence labels
don't match between the alignments:

  individual-1-locus-1-pop-1  ATT???
  individual-2-locus-1-pop-1  ATT???
  individual-1-locus-1-pop-2  TTT???
  individual-2-locus-1-pop-2  ATA???
  individual-1-locus-2-pop-1  ???GGC
  individual-2-locus-2-pop-1  ???GGC
  individual-1-locus-2-pop-2  ???GGA
  individual-2-locus-2-pop-2  ???TGC

The two loci end up disjoint, but this won't affect how BiallelicData
processes the data. In fact, the test added confirms that every site can
be disjoint. This makes sense because the BiallelicData assumes every
site is independent.