Fixing bug in parsing of pop size multipliers.
Recent changes in the PLL branch to how the settings module parses and
handles PositiveRealParameterSettings that store a vector of values
introduced a bug in the parsing of 'population_size_multiplers' in
DirichletComparisonSettings. We were attempting to normalize the empty
vector of values when they were not specified in the config file.
Fixing the bug by adding a check to make sure the values are not empty
(i.e., ensuring the values were specified in the yaml config) before
normalizing.